omap: clock: not all clocks have .round_rate
Protect clk_set_rate function from the case where a clock does
not have a round_rate function pointer.  In such a case we simply
continue with the assumption that the rate passed into clk_set_rate
is OK.

This seems a bit dubious but if a clock does not have a round_rate
it is assumed it also does not have a set_rate function and will be
handled properly by clk_set_rate.

Change-Id: Ib0cce595fe42bbb450cb91196727b14823900bfe
Signed-off-by: Mike Turquette <mturquette@ti.com>
Signed-off-by: Margarita Olaya <magi.olaya@ti.com>