omap: uart: wer restoration
The WER value for a uart port is restored properly in
either case
Case 1:
The port is started, the device enters off mode and
resumes. In this case wer is restored with the
intended board defined value.

Case 2:
The port is started, shutdown, the device enters off
mode and resumes. In this case wer is restored with zero.
This prevents events from being registered in the uart
hardware which will not be handled by a port that is not
started in software.

Change-Id: I5dd426b2cdb8925404f6ba7fc83569b49bda1695
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>