PVR: sgxfreq: Introduce function to get GPU load in sgxfreq.c
Previously, load calculations were performed in on3demand. Move
this function to sgxfreq.c, along with the sysfs parameter so it
is available to all governors that wish to use it. This opens up
the possibility of creating new governors that rely on load
calculations without having to introduce redundant code.

Note: GPU load will display as 0 if the governor does not request it.